restructuring and update for 2023
instead of having the current year in / and the archived ones
moved to /archive/YYYY, it seems better to just have /YYYY.
Let / redirect to the current year. And add deprecation markers to
a year once it's not the current year anymore.

Added redirects where applicable so people browsing to the old pages
still get the same content.

Added teaser for 2023